requirements.txt
alembic==1.7.7
APScheduler==3.10.1
# backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==2.0.12
click==8.0.4
dataclasses
decorator==5.1.1
Flask==2.0.3
Flask-Migrate==4.0.1
Flask-SQLAlchemy==2.5.1
google-search-results==2.4.2
greenlet==2.0.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mysqlclient==2.1.1

python-dotenv==0.20.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.48
typing-extensions==4.1.1
tzdata==2023.3
tzlocal==4.2
urllib3==1.26.16
validators==0.20.0
Werkzeug==2.0.3
zipp==3.6.0
